CALM-Continual_Adaptive_Learning_Model/requirements.txt
torch==2.3.1
torchvision==0.18.1
torchaudio==2.3.1
transformers==4.41.2
Pillow==10.3.0
scikit-learn==1.5.0